+++ imports/api/aws/index.js
@@ -7,7 +7,9 @@ if (Meteor.isServer) {
   Meteor.methods({
     'aws.write'(dataUrl, filePath) {
       var buff = new Buffer(dataUrl.split(',')[1], 'base64');
-      var s3 = new AWS.S3({apiVersion: '2006-03-01'});
+      var s3 = new AWS.S3({
+        apiVersion: '2006-03-01'
+      });
       if (Meteor.isDevelopment) {
         filePath = "tests/" + filePath;
       }
@@ -18,6 +20,7 @@ if (Meteor.isServer) {
         Key: filePath
       }
       return new Promise((resolve, reject) => {
+        console.log(s3.credentials);
         s3.upload(params, (err, data) => {
           if (err) {
             reject(err);